Train (live) data, from db, to facade, to frontend (needs work)

+++ backend/hermes/src/main/java/com/trinity/hermes/indicators/train/repository/TrainStationRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TrainStationRepository extends JpaRepository<TrainStation, Long> {
+public interface TrainStationRepository extends JpaRepository<TrainStation, Integer> {
   TrainStation findByStationCode(String stationCode);
 }
